+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Healthcheck `kind` has been renamed to `type` and it's values are now lowercase (eg. `type: exec`).
 
+### Fixed
+- Paths within module definitions are now correctly resolved relative to the client rather than the daemon.
+
 ## [0.3.0-alpha] - 2020-01-11
 ### Added
 - Added a new `Group` declaration to allow for modules whose sole purpose is to group other modules. They can be deployed, specified as a dependency, or have checks just like other modules.